requirements.txt
altgraph==0.16.1
astroid==2.3.3
azure-devops==5.0.0b9
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
et-xmlfile==1.0.1
future==0.17.1
idna==2.8
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
macholib==1.11
mccabe==0.6.1
msrest==0.6.8
oauthlib==3.1.0
openpyxl==2.6.2
pefile==2019.4.18
PyInstaller==3.5
pylint==2.4.4
pywin32-ctypes==0.2.0
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
tqdm==4.39.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
